pyproject.toml
[tool.poetry]
name = "benchmark-demo"
version = "0.1.0"
description = "A benchmark for testing tf methods."
authors = ["Juan Manuel Miramont-Taurel <jmiramontt@univ-lille.fr>"]
packages = [{include = "src"}]

# Basic dependencies
[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.22.0"
matplotlib = "^3.5.1"
pandas = "^1.4.4"
scipy = "^1.7.3"
seaborn = "^0.12.0"
cmocean = "^2.0"
pytest = "^7.0.1"
tabulate = "^0.8.9"
PyYAML = "^6.0"
plotly = "^5.10.0"
typeguard = "^2.13.3"
librosa = "^0.9.2"
mcsm-benchs = {git = "https://github.com/jmiramont/mcsm-benchs.git"}

# Install R-based tools with "poetry install --with rtools"
[tool.poetry.group.rtools]
optional = true

[tool.poetry.group.rtools.dependencies]
rpy2 = "^3.5.4"
spatstat-interface = "^1.0.1"

# Install Matlab-based tools with "poetry install --with matlab_tools"
[tool.poetry.group.matlab_tools]
optional = true

[tool.poetry.group.matlab_tools.dependencies]
matlabengine = "9.13.6"

# Install Notebooks dependencies "poetry install --with notebooks"
[tool.poetry.group.notebooks]
optional = true

[tool.poetry.group.notebooks.dependencies]
ipykernel = "^6.25.2"

# Install Docs dependencies "poetry install --with docs"
[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
Sphinx = {version = ">=6,<8"}
sphinx-rtd-theme = {version = "^1.0.0"}
nbsphinx = { version = "^0.9.2"}
sphinxcontrib-bibtex = { version = "^2.5.0"}
myst-parser = { version = "^2.0.0"}
sphinx-copybutton = { version = "^0.5.2"}
sphinx-sizzle-theme = "^0.1.3"
sphinx-material = "^0.0.36"
# sphinx-material = {git ="https://github.com/bashtage/sphinx-material.git"}

# Development dependencies
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
